VW-936 Change datetype requestOutsource, outsourceOrder

+++ Infrastructure/EntityFramework/ManufacturingDB/OutsourceOrderDetail.cs
@@ -5,8 +5,8 @@ namespace VErp.Infrastructure.EF.ManufacturingDB
 {
     public partial class OutsourceOrderDetail
     {
-        public int OutsourceOrderDetailId { get; set; }
-        public int OutsoureOrderId { get; set; }
+        public long OutsourceOrderDetailId { get; set; }
+        public long OutsoureOrderId { get; set; }
         public int ObjectId { get; set; }
         public decimal Price { get; set; }
         public decimal Tax { get; set; }
