+++ Service.Manafacturing.Model/Outsource/Order/OutsourceOrderDetailModel.cs
@@ -8,8 +8,8 @@ namespace VErp.Services.Manafacturing.Model.Outsource.Order
 {
     public class OutsourceOrderDetailModel: IMapFrom<OutsourceOrderDetail>
     {
-        public int OutsourceOrderDetailId { get; set; }
-        public int OutsoureOrderId { get; set; }
+        public long OutsourceOrderDetailId { get; set; }
+        public long OutsoureOrderId { get; set; }
         public int ObjectId { get; set; }
         public decimal Price { get; set; }
         public decimal Tax { get; set; }
